chore: rename block_explorer.ts to blockExplorer.ts for consistency
